requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.0
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Geospatial
geopandas==0.14.2
shapely==2.0.2
pyproj==3.6.1
geoalchemy2==0.14.3
numpy==1.26.4

# HTTP client
httpx[http2]==0.27.0
tenacity==8.2.3

# Caching
aiocache==0.12.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
authlib>=1.3.0
itsdangerous>=2.1.2

# Encryption
cryptography>=41.0.0

# File handling
aiofiles==23.2.1

# CORS
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.27.0
aiosqlite==0.19.0

# MCP Server
mcp>=1.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
